+++ bg_3D.cpp
@@ -17,8 +17,8 @@
 //***************************
 //’è”‚Ì’è‹`
 //***************************
-const float CBg3D::BG_WIDTH = CRenderer::SCREEN_WIDTH;		//‰¡•
-const float CBg3D::BG_HEIGHT = CRenderer::SCREEN_HEIGHT;	//c•
+const float CBg3D::BG_WIDTH = (float)CRenderer::SCREEN_WIDTH;	//‰¡•
+const float CBg3D::BG_HEIGHT = (float)CRenderer::SCREEN_HEIGHT;	//c•
 
 //================================================
 //¶¬
